Preserve teams across scenes
This mostly adds PersistentPlayersManager, which will keep track of devices (differentiated by PlayerInput.playerIndex) and the corresponding PlayerID string.

This turned out to not be necessary for keeping teams the same in each level, for now - in GameManager, the team is now just determined by the PlayerInput playerIndex instead of current player count.  This should definitely be replaced by a proper team select screen.