feat: add hat movement to ghost player
The hat at this point statically moves around mirroring the ghost wiggle
speed. What would be even better is to have it rotate slightly with
the ghost head.

If we want "soft" hats, I could also make another shader that wiggles
like the ghost, but on top of the ghost head.