Move TeamData instances out of each PlayerData, and into its own array/dict
Now PlayerData just stores a TeamData.Team enum, which is used to fetch the actual TeamData from Persistent Players Manager.

Also noticed while doing this - PersistentPlayersManager is not moved to Don'tDestroyOnLoad unless it is a root gameobject!  Might need to make it not do that.

Main menu change just makes the PersistentPlayersManager there a prefab.