Add togglable keyboard control support, via the DebugManager
Instead of turning on keyboard support by default for players, this adds a way to turn it on manually for individual players.  Hit the F10 key to cycle between all players (a debug log will print the current one) and you will be able to control that player with the keyboard.