Generic hat possessable prefab + hat selection
Now the Hat Possessable prefab can be used for all hats - setting its HatData will instantiate the correct mesh/collision.  Players can possess it (to drag it around), then haunt it to put the hat on.  Putting on a new hat causes the old hat to pop off as a possessable.

Not sure if the hat data is preserved between levels yet - have to check that later.